refactor(Video): remove interface
