+++ src/service/VideoService.java
@@ -1,10 +1,22 @@
 package service;
 
 import model.Video;
+import repository.VideoRepository;
 
 import java.util.List;
 
-public interface VideoService {
-    void addVideo(Video video);
-    List<Video> listVideos();
+public class VideoService {
+    private final VideoRepository repository;
+
+    public VideoService(VideoRepository repository) {
+        this.repository = repository;
+    }
+
+    public void addVideo(Video video) {
+        repository.save(video);
+    }
+
+    public List<Video> listVideos() {
+        return repository.findAll();
+    }
 }
