--- src/service/VideoServiceImpl.java
@@ -1,24 +1,0 @@
-package service;
-
-import model.Video;
-import repository.VideoRepository;
-
-import java.util.List;
-
-public class VideoServiceImpl implements VideoService {
-    private final VideoRepository repository;
-
-    public VideoServiceImpl(VideoRepository repository) {
-        this.repository = repository;
-    }
-
-    @Override
-    public void addVideo(Video video) {
-        repository.save(video);
-    }
-
-    @Override
-    public List<Video> listVideos() {
-        return repository.findAll();
-    }
-}
